+++ PSoC/SensorSender/SensorSender.cydsn/sensor.h
@@ -2,5 +2,5 @@ void initSensor();
 void startBurst();
 void stopBurst();
 void startCounter();
-double calcTime(double count);
-double calcAngle(long count);
+float calcTime(long count);
+float calcAngle(long count);
